backend/pyproject.toml
[tool.poetry]
name = "vote"
version = "1.4.1"
description = "Vote"
authors = ["Osayd Abdu <OsaydAbdu@gmail.com>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "~3.9"
tenacity = "^6.1.0"
typing-extensions = "^4.0.1"
uvicorn = "^0.17.4"
fastapi = "^0.68.1"
gunicorn = "^20.1.0"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
email-validator = "^1.1.3"
password-strength = "^0.0.3.post2"
sqlalchemy = "^1.3.16"
psycopg2 = "^2.9.3"
alembic = "^1.7.5"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-watch = "^4.2.0"
pytest-cov = "^2.12.1"
cython = "^0.29.24"
flake8 = "^3.9.2"
isort = "^5.9.3"


[tool.isort]
profile="black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88



[build-system]
requires = ["poetry>=0.12", "poetry-core>=1.0.0"]
build-backend = "poetry.masonry.api"
